feature: Generate signed URLs for blobs (#1)
Add pre-signed URL generation for AWS, GCP and Azure. Pre-signed URLs
don't make much sense for in-memory or filesystem backends, so they are
not implemented.

Note: this deliberately only tests AWS S3, because that's
straightforward with `moto`; it includes TODOs for testing with Azure,
but as I don't have a need for it it remains undone. There isn't a good
way to test with GCP except with real credentials, so that is also
commented out in the tests.